requirements.txt
aiodns==3.0.0
aiohttp==3.7.4.post0
astroid==2.7.3
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==21.2.0
Authlib==0.15.4
bcrypt==3.2.0
cachetools==4.2.2
cchardet==2.1.7
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.3
click==7.1.2
colorama==0.4.4
coverage==5.5
cryptography==3.4.7
decorator==5.0.9
elasticsearch==7.0.0
Flask==1.1.4
Flask-Bcrypt==0.7.1
Flask-Caching==1.10.1
gunicorn==20.1.0
idna==2.10
iniconfig==1.1.1
install==1.3.4
isort==5.9.3
itsdangerous==1.1.0
Jinja2==2.11.3
lark-parser==0.11.3
lazy-object-proxy==1.6.0
libcst==0.3.19
MarkupSafe==2.0.1
mccabe==0.6.1
multidict==5.1.0
mypy-extensions==0.4.3
packaging==21.0
pika==1.2.0
pika-stubs==0.1.3
platformdirs==2.3.0
pluggy==0.13.1
proto-plus==1.19.0
protobuf==3.17.3
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==4.0.0
pycparser==2.20
PyJWT==2.1.0
pymemcache==3.5.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-decouple==3.4
python3-pika==0.9.14
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-toolbelt==0.9.1
required==0.4.0
retry==0.9.2
rsa==4.7.2
schedule==1.1.0
seconds==0.1.0
shortuuid==1.0.1
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.0
typing-inspect==0.7.1
tzlocal==2.1
urllib3==1.26.6
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.6.3